ID-1561 Example app with custom plugin class using dependency injection
The app 'app-with-extensions' provides a custom plugin class which makes
use of dependency injection to register a custom helper for showing the
current date and time. This is shown in example page ex:AppExtensions.

To avoid conflicts with other app (e.g. wikidata) the prefix namespace
'app:' has been remapped to 'ex:' also in app ',y-first-app' and is
consistent with the definition in 'app-with-extensions'.