Switch to SDK bundled in metaphactory (automated setup via prepareEnvironment script)
This PR removes the SDK as it was moved to the metaphactory distribution. Instead, the SDK can be retrieved from the docker image by running the included prepareEnvironment script. This script pulls the docker image, extracts the platform WAR file and SDK and sets everything up properly.

Usage of prepareEnvironment.sh script:

-i <image> # run specific docker image
-f force update no matter if SDK already exists or if it's the same version